requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
dtreeviz
Pillow
graphviz
